"Bootstrap 5 Migration - initial auto-migration for app_manager/js/nav_menu_media_common.js, splitting templates"
